pyproject.toml
[project]
name = "calendar-workload-runner"
version = "0.1.0"
description = "Google Calendar based workload runner for WSL"
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.12"
license = { file = "LICENSE" }
authors = [{ name = "Kiminori KATO" }]

dependencies = [
    "google-api-python-client",
    "google-auth-httplib2",
    "google-auth-oauthlib",
    "python-dateutil",
]

classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/kimikato/calendar-workload-runner"
Source = "https://github.com/kimikato/calendar-workload-runner"
Issues = "https://github.com/kimikato/calendar-workload-runner/issues"
Documentation = "https://github.com/kimikato/calendar-workload-runner/tree/main/docs"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "mypy>=1.9",
    "pytest>=8.0",
    "coverage",
    "ruff>=0.4",
    "build",
    "types-python-dateutil",
]

[project.scripts]
calendar_workload_runner = "calendar_workload_runner.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

# ---------------------------------------------------------
# Pytest
# ---------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

# ---------------------------------------------------------
# Black
# ---------------------------------------------------------
[tool.black]
line-length = 100
target-version = ["py312"]
skip-string-normalization = true


# ---------------------------------------------------------
# Mypy (strict mode)
# ---------------------------------------------------------
[tool.mypy]
python_version = "3.12"
mypy_path = "src"
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
no_implicit_optional = true
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_any_generics = true
show_error_codes = true
show_error_context = true
show_column_numbers = true
pretty = true

[[tool.mypy.overrides]]
module = [
    "googleapiclient.*",
    "google_auth_oauthlib.*",
    "google.oauth2.*",
    "google.auth.*",
]
ignore_missing_imports = true


# ---------------------------------------------------------
# Coverage
# ---------------------------------------------------------
[tool.coverage.run]
branch = true
source = ["src/calendar_workload_runner"]

[tool.coverage.report]
show_missing = true
skip_covered = true
